requirements/requirements.txt
opencv-python==4.9.0.80
scikit-build==0.17.6
pandas==2.0.3
tensorboard==2.14.0
vit-pytorch==0.20.0
tqdm==4.66.1
ipywidgets==8.1.1
timm==0.9.12
scipy==1.10.1
patool==1.12
setuptools==59.5.0
scikit-learn==1.3.2